+++ src/eni/seis/content/product_item.py
@@ -8,12 +8,17 @@ class ProductItem(Container):
     """ ProductItem content type """
 
     def get_view_url(self):
-        """ Return its external link if exists or absolute_url
+        """ Return its external link or file download url if exists
+            OR absolute_url
             (because in some cases a report can act like a page with links or
             some other content)
         """
+        if self.has_file():
+            return self.absolute_url() + "/@@download/file"
+
         if self.has_external_link():
             return self.external_link
+
         return self.absolute_url()
 
     def has_external_link(self):